refactor(fonts,terminal): extract font manager and terminal manager into workspace crates (#165, #166)
- Extract FontManager, TextShaper, and fallback chain into par-term-fonts crate
- Extract TerminalManager, scrollback metadata, styled content, and terminal
  submodules into par-term-terminal crate
- Move Cell type to par-term-config as shared type between terminal and renderer
- Add integration tests for font loading, text shaping, and terminal lifecycle
- Main crate re-exports all types for backward compatibility
