+++ par-term-fonts/Cargo.toml
@@ -1,0 +1,36 @@
+[package]
+name = "par-term-fonts"
+version = "0.1.0"
+edition = "2024"
+authors = ["Paul Robello <probello@gmail.com>"]
+description = "Font management and text shaping for par-term terminal emulator"
+license = "MIT"
+repository = "https://github.com/paulrobello/par-term"
+homepage = "https://github.com/paulrobello/par-term"
+keywords = ["terminal", "font", "text-shaping"]
+categories = ["text-processing"]
+
+[dependencies]
+# Font rasterization with colored glyph support
+swash = "0.2.6"
+
+# Font database for system font discovery and fallback
+fontdb = "0.23"
+
+# HarfBuzz text shaping for ligatures and complex scripts
+rustybuzz = "0.20.1"
+
+# Grapheme cluster detection for proper text boundaries
+unicode-segmentation = "1.12"
+
+# LRU cache for text shaping results
+lru = "0.16.3"
+
+# Configuration types (FontRange)
+par-term-config = { path = "../par-term-config" }
+
+# Error handling
+anyhow = "1.0.101"
+
+# Logging
+log = "0.4.29"
