+++ par-term-terminal/Cargo.toml
@@ -1,0 +1,36 @@
+[package]
+name = "par-term-terminal"
+version = "0.1.0"
+edition = "2024"
+authors = ["Paul Robello <probello@gmail.com>"]
+description = "Terminal manager for par-term terminal emulator"
+license = "MIT"
+repository = "https://github.com/paulrobello/par-term"
+homepage = "https://github.com/paulrobello/par-term"
+keywords = ["terminal", "pty", "emulator"]
+categories = ["command-line-utilities"]
+
+[dependencies]
+# Terminal emulator core
+par-term-emu-core-rust = { version = "0.39.0", default-features = false, features = ["rust-only"] }
+
+# Configuration (for Theme, Color, Cell types)
+par-term-config = { path = "../par-term-config" }
+
+# Synchronization
+parking_lot = "0.12.5"
+
+# Async runtime (for paste_with_delay)
+tokio = { version = "1.49", features = ["time", "rt"] }
+
+# Error handling
+anyhow = "1.0.101"
+
+# Logging
+log = "0.4.29"
+
+# Clipboard support
+arboard = "3.6.1"
+
+[dev-dependencies]
+tempfile = "3.25"
