+++ par-term-terminal/src/terminal/clipboard.rs
