+++ src/font_manager.rs
@@ -1,0 +1,6 @@
+//! Font management with fallback chain for comprehensive Unicode coverage.
+//!
+//! This module re-exports types from the par-term-fonts crate for backward compatibility.
+//! All font management types, loading, and text shaping are defined in par-term-fonts.
+
+pub use par_term_fonts::font_manager::*;
